Rework config diff infra
Instead of diffing properties and groups separately we know diff them
in a consistent way. In theory, this is not 100% correct since the
properties collection returned from the group is a merged version in
case several properties exist within the same group. This is probably
the case with spring.datasource but we can ignore that problem for now.

Note that property diff is not implemented yet.