Use working dir rather than hard-coded value
Update the static locations to use the `USER_DIR` environment variable
that can be accessed via ${user.dir} in configuration.

This assumes that the application runs from the root of the project. It
is possible to configure Intellij IDEA to do so by specifying the
`$PROJECT_DIR$` macro in the Working Directory field of the run
configuration.