Enable store by reference
By default, JSR-107 stores items by value. When DevTools is enabled, a
cache hit leads to the deserialization of the item from the cache. If
that deserialization does not happen in the same `Classloader` than the
one the application uses, we get a `ClassCastException` as the item's
class is not equal to the one provided by the current classloader.

EhCache is using the JAR classloader by default. To workaround that we
enable the store by reference instead.