Metrics demonstration
Once the actuator is in place, this is very easy to augment some of the
endpoints to expose application-specific meta-data.

Metrics can be easily recorded (both individual values and counters). In
this commit, we record the number of times a particular version was
requested. We also use the new `@ExportMetricWriter` annotation to
export our metrics to a separate system. In this case, we simply expose
them to JMX but plenty of other options are available (redis, statsd,
etc).